+++ src/main/java/ro/ciprian/curs/JavaApplication.java
@@ -1,4 +1,4 @@
-package ro.andu.curs;
+package ro.ciprian.curs;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
