+++ src/main/java/ro/ciprian/curs/configuration/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.configuration;
+package ro.ciprian.curs.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
