+++ src/main/java/ro/ciprian/curs/configuration/ciprian/CarRepositoryConfigCiprian.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.configuration.andu;
+package ro.ciprian.curs.configuration.ciprian;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
@@ -8,24 +8,24 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
-import ro.andu.curs.model.Car;
-import ro.andu.curs.repository.MemoryRepository;
+import ro.ciprian.curs.model.Car;
+import ro.ciprian.curs.repository.MemoryRepository;
 
 @Configuration
-@Profile("andu")
-public class CarRepositoryConfigAndu {
+@Profile("ciprian")
+public class CarRepositoryConfigCiprian {
     @Bean
     public void loadDataDev() {
         List<Car> list = new ArrayList<>();
 
         Car car1 = new Car();
-        car1.setMaker("Ford");
-        car1.setColor("black");
-        car1.setModel("Mustang");
+        car1.setMaker("Audi");
+        car1.setColor("green");
+        car1.setModel("A8");
         car1.setYear(2023);
         car1.setCurrency("EUR");
         car1.setMileage(1000);
-        car1.setPrice(BigDecimal.valueOf(100000));
+        car1.setPrice(BigDecimal.valueOf(1727420));
         list.add(car1);
 
         MemoryRepository.carList=list;
