+++ src/main/java/ro/ciprian/curs/controller/v1/HelloWorldController.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.controller.v1;
+package ro.ciprian.curs.controller.v1;
 
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import ro.andu.curs.dto.HelloRequest;
+import ro.ciprian.curs.dto.HelloRequest;
 
 @RestController
 @RequestMapping("/v1")
