+++ src/main/java/ro/ciprian/curs/dto/SedanDto.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.dto;
+package ro.ciprian.curs.dto;
 
 public class SedanDto extends CarDto{
     private Integer doors;
