+++ src/main/java/ro/ciprian/curs/dto/SuvDto.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.dto;
+package ro.ciprian.curs.dto;
 
 public class SuvDto extends CarDto{
     private Integer transmission;
