+++ src/main/java/ro/ciprian/curs/dto/VanDto.java
@@ -1,4 +1,4 @@
-package ro.andu.curs.dto;
+package ro.ciprian.curs.dto;
 
 public class VanDto extends CarDto {
     private String capacity;
