+++ src/main/java/ro/ciprian/curs/repository/CarRepository.java
@@ -1,9 +1,10 @@
-package ro.andu.curs.repository;
+package ro.ciprian.curs.repository;
 
 import org.springframework.stereotype.Repository;
 
+import ro.ciprian.curs.model.Car;
+
 import java.util.List;
-import ro.andu.curs.model.Car;
 
 @Repository
 public class CarRepository {
